Move paint function into utility
