+++ tests/unit/utils/build-mapbox-paint-object-test.js
@@ -1,0 +1,10 @@
+import buildMapboxPaintObject from 'labs-regional-viz/utils/build-mapbox-paint-object';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | build mapbox paint object');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = buildMapboxPaintObject();
+  assert.ok(result);
+});
